+++ lib/timex.ex
@@ -54,6 +54,12 @@ defmodule Timex do
   defdelegate today(), to: Date, as: :utc_today
 
   @doc """
+  Returns a Date representing the current day in the provided timezone.
+  """
+  @spec today(Types.valid_timezone()) :: Date.t()
+  def today(timezone), do: now(timezone) |> DateTime.to_date()
+
+  @doc """
   Returns a DateTime representing the current moment in time in UTC
   """
   @spec now() :: DateTime.t()